PR fixes, refactoring files
