+++ src/Components/Users/UsersForm.js
@@ -1,73 +1,11 @@
-import React, { useState, useRef } from "react";
-import { Row, Spinner} from "react-bootstrap";
-import { ErrorMessage, Field, Form, Formik,   } from "formik";
-import * as Yup from "yup";
-import { toBase64 } from "../../Helpers/base64";
-import { Patch, Post } from "../../Services/privateApiService";
+import React from "react";
+import { Row, Spinner } from "react-bootstrap";
+import { ErrorMessage, Field, Form } from "formik";
 import "../FormStyles.css";
 import "bootstrap/dist/css/bootstrap.min.css";
-import { NETWORK_ERROR, UNKNOWN_ERROR, EMAIL_TAKEN } from "../../Helpers/messagesText";
-import { yupTitles ,yupEmail, yupShortDesc, yupImages, yupUserRoles, yupPassword,  } from "../../Helpers/yupValidations";
 
-const validation = Yup.object().shape({
-  name: yupTitles(),
-  email: yupEmail(),
-  description: yupShortDesc(),
-  image_file: yupImages(),
-  role_id: yupUserRoles(),
-  password: yupPassword()
-});
+import userTypes from "../../Helpers/userTypes";
 
-const initialValues = {
-  name: "",
-  email: "",
-  description: "",
-  role_id: 2,
-  image_file: null,
-  profile_image:'',
-  password: ""
-};
-
-const submit = async (
-  values,
-  setSubmitting,
-  user,
-  setSuccess,
-  setRequestError,
-  resetForm,
-  fileInputRef
-) => {
-  
-  setRequestError("");
-  values.role_id = Number(values.role_id);
-  
-  values.profile_image = await toBase64(values.image_file)
-
-  try {
-    const response = user ? await Patch(`users/${user.id}`, values) : await Post(`users`, values);
-    
-    if (response.data.success) {
-      setSuccess(true);
-      resetForm();
-      fileInputRef.current.value = null
-    } else {
-      setRequestError(UNKNOWN_ERROR);
-    }
-
-    setSubmitting(false);
-
-  } catch (error) {
-    if (error.message === "Network Error") {
-      setRequestError(NETWORK_ERROR);
-    } else if (error.response.data.errors.email) {
-      setRequestError(EMAIL_TAKEN)
-    } else {
-      setRequestError(UNKNOWN_ERROR);
-    }
-
-    setSubmitting(false);
-  }
-};
 
 const FormComponent = ({
   touched,
@@ -76,7 +14,7 @@ const FormComponent = ({
   success,
   requestError,
   isSubmitting,
-  fileInputRef
+  fileInputRef,
 }) => {
   return (
     <Form onSubmit={handleSubmit} className="form-container">
@@ -117,8 +55,9 @@ const FormComponent = ({
       {touched.password && <ErrorMessage name="password" />}
 
       <Field className="select-field" as="select" name="role_id">
-        <option value='2'>Usuario Standard</option>
-        <option value='1'>Administrador</option>
+        {Object.keys(userTypes).map((key)=>{
+          return <option value={key}>{userTypes[key]}</option>
+        })}
       </Field>
       <div>
         <label style={{ paddingRight: "10px" }}>Subir imagen de usuario</label>
@@ -153,43 +92,4 @@ const FormComponent = ({
   );
 };
 
-const UserForm = ({ user }) => {
-  const [success, setSuccess] = useState(false);
-  const [requestError, setRequestError] = useState();
-
-  const fileInputRef = useRef()
-
-  return (
-    <Formik
-      validationSchema={validation}
-      initialValues={user || initialValues}
-      onSubmit={(values, { setSubmitting, resetForm }) =>
-        submit(
-          values,
-          setSubmitting,
-          user,
-          setSuccess,
-          setRequestError,
-          resetForm,
-          fileInputRef
-        )
-      }
-    >
-      {({ handleSubmit, setFieldValue, touched, isSubmitting }) => {
-        return (
-          <FormComponent
-            touched={touched}
-            handleSubmit={handleSubmit}
-            setFieldValue={setFieldValue}
-            success={success}
-            requestError={requestError}
-            isSubmitting={isSubmitting}
-            fileInputRef={fileInputRef}
-          />
-        );
-      }}
-    </Formik>
-  );
-};
-
-export default UserForm;
+export default FormComponent
