+++ src/Helpers/formValidations.js
@@ -1,5 +1,14 @@
 import * as Yup from "yup";
-import {TITLE_LONG, TITLE_SHORT, INPUT_REQUIRED, DESCRIPTION_LONG, DESCRIPTION_SHORT, INVALID_IMAGE_EXT, INVALID_EMAIL, PASSWORD_SHORT} from './messagesText'
+import {
+  TITLE_LONG,
+  TITLE_SHORT,
+  INPUT_REQUIRED,
+  DESCRIPTION_LONG,
+  DESCRIPTION_SHORT,
+  INVALID_IMAGE_EXT,
+  INVALID_EMAIL,
+  PASSWORD_SHORT,
+} from "./messagesText";
 
 const yupTitles = () =>
   Yup.string()
@@ -28,25 +37,25 @@ const yupCustomString = (min, max, minMsg, maxMsg) =>
 const yupImages = () =>
   Yup.mixed()
     .required(INPUT_REQUIRED)
-    .test(
-      "fileFormat",
-      INVALID_IMAGE_EXT,
-      (value) => ["image/jpeg", "image/png"].includes(value?.type)
+    .test("fileFormat", INVALID_IMAGE_EXT, (value) =>
+      ["image/jpeg", "image/png"].includes(value?.type)
     );
 
-const yupEmail = () =>   
-  Yup.string()
-    .email(INVALID_EMAIL)
-    .required(INPUT_REQUIRED)
+const yupEmail = () =>
+  Yup.string().email(INVALID_EMAIL).required(INPUT_REQUIRED);
 
-const yupUserRoles = () => 
-  Yup.string()
-    .required(INPUT_REQUIRED)
+const yupUserRoles = () => Yup.string().required(INPUT_REQUIRED);
 
 const yupPassword = () =>
-  Yup.string()
-    .required(INPUT_REQUIRED)
-    .min(8, PASSWORD_SHORT)
+  Yup.string().required(INPUT_REQUIRED).min(8, PASSWORD_SHORT);
 
-
-export {yupCustomString, yupEmail, yupImages, yupLongDesc, yupPassword, yupShortDesc, yupTitles, yupUserRoles}
+export {
+  yupCustomString,
+  yupEmail,
+  yupImages,
+  yupLongDesc,
+  yupPassword,
+  yupShortDesc,
+  yupTitles,
+  yupUserRoles,
+};
