+++ src/Helpers/userTypes.js
@@ -1,0 +1,1 @@
+export default { 2: "Usuario Standard", 1: "Usuario Adminisrador" };
